requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
asyncpg
python-dotenv
pydantic
psycopg2-binary
oauthlib
yandex-oauth
passlib[bcrypt]
python-jose[cryptography]
pytz